sanitize: use PascalCase for enums, switch from `inflections` to `convert_case`.
