+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.5.0", features = ["derive"] }
 env_logger = "0.11.1"
-inflections = "1.1"
+convert_case = "0.11"
 log = { version = "~0.4", features = ["std"] }
 quote = "1.0"
 proc-macro2 = "1.0"
